refactor import functions out backend compiler, fix references and fix test failing due to var change.
