+++ ivy/functional/backends/numpy/__init__.py
@@ -12,8 +12,8 @@ if not ivy.is_local():
     _module_in_memory = sys.modules[__name__]
 else:
     # TODO remove str dependency
-    global_backend_compiler = sys.modules["ivy.utils.backend.compiler"]
-    _module_in_memory = global_backend_compiler.IMPORT_CACHE[__name__]
+    global_backend_compiler = sys.modules["ivy.utils.backend._importlib"]
+    _module_in_memory = global_backend_compiler.import_cache[__name__]
 
 use = ivy.utils.backend.ContextManager(_module_in_memory)
